assignment-10/requirements.txt
pinecone-client>=6.0.0
openai>=1.99.0
PyPDF2>=3.0.0
